[FEAT] Input text is split into separate htmls
The assumption here is that the input is a single
file with one or more well-formed html files.
Each of them has a !DOCTYPE header, so it is
used to do the split. If it is a single
document, it does not matter; if it is a single
document with no header at all -- it does not
matter either as there will be nothing to
split it into.